% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSG-package.R
\docType{data}
\name{cn_vs_us}
\alias{cn_vs_us}
\title{Country power indicators of China vs America}
\description{
Country power indicators of China vs America
}
\examples{
data(cn_vs_us)
}
\references{
\url{https://www.theguardian.com/news/datablog/2011/jan/19/china-social-media}
}
