% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSG-package.R
\docType{data}
\name{eq2010}
\alias{eq2010}
\title{Longitude and latitude of earthquakes in the Sichuan Province}
\description{
Longitude and latitude of earthquakes in the Sichuan Province
}
\examples{
data(eq2010)
plot(lat ~ long, data = eq2010)
}
