\name{inSide}
\alias{inSide}
\title{Are points inside a polygon?}
\usage{
  inSide(bnd, x, y)
}
\arguments{
  \item{bnd}{a polygon boundary (a list with elements x and
  y)}

  \item{x}{x coordinates of the points to test}

  \item{y}{y coordinates of the points to test}
}
\value{
  logical vector
}
\description{
  This is Simon Wood's inSide from soap
}

