\name{MSG-package}
\alias{MSG-package}
\alias{MSG-package}
\title{Modern Statistical Graphics}
\description{
Datasets and functions for the Chinese book ``Modern Statistical Graphics''.}
\details{\tabular{ll}{ Package: \tab MSG\cr Type: \tab Package\cr License:
\tab GPL\cr LazyLoad: \tab yes\cr }}
\alias{MSG-package}
\alias{MSG}
\docType{package}
\author{Yihui Xie <xie@yihui.name>}
\keyword{package}
