% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msg.R
\name{msg}
\alias{msg}
\title{Plot a graph with a pre-installed R script}
\usage{
msg(fig = 3.6, show_code = TRUE)
}
\arguments{
\item{fig}{The figure number or the R script name, which is given in the book.}

\item{show_code}{Logical.}
}
\value{
A graph and the source code
}
\description{
Plot a graph with a pre-installed R script
}
\examples{
# msg(3.6) msg('ChinaPop')
}
