\name{Predict.matrix.dm.smooth}
\alias{Predict.matrix.dm.smooth}
\title{Distance matrix smoothing}
\usage{
  \method{Predict.matrix}{dm.smooth} (object, data)
}
\arguments{
  \item{object}{a smooth specification object, usually
  generated by a term \code{s(...,bs="ds",...)}. Note that
  \code{xt} object is needed, see Details.}

  \item{data}{a list containing just the data (including
  any \code{by} variable) required by this term, with names
  corresponding to \code{object$term} (and
  \code{object$by}). The \code{by} variable is the last
  element.}
}
\description{
  Provides smoothing methods for multidimensional
  scaling-based projections of a given distance matrix.
}
\details{
  Smoothing is performed using Duchon splines (see
  \code{\link{Duchon.spline}} for more information).
}

