\name{Predict.matrix.msg.smooth}
\alias{Predict.matrix.msg.smooth}
\title{Multidimensional Scaling for Generalized additive models - msg}
\usage{
  \method{Predict.matrix}{msg.smooth} (object, data)
}
\arguments{
  \item{object}{a smooth specification object, usually
  generated by a term \code{s(...,bs="ds",...)}. Note that
  \code{xt} object is needed, see Details.}

  \item{data}{a list containing just the data (including
  any \code{by} variable) required by this term, with names
  corresponding to \code{object$term} (and
  \code{object$by}). The \code{by} variable is the last
  element.}
}
\description{
  Provides smoothing methods for multidimensional
  scaling-based projections of the data.
}
\details{
  Usage is split into two cases: (1) For geographical
  smooths of two coordinates, within-area distances with
  respect to the boundary are used to create a distance
  matrix that is then projected into as many dimensions as
  required. (2) General distance smoothing where distances
  are calculated  between observations based on all
  covariates supplied using Euclidean or Mahalanobis
  distances.

  In both cases smoothing is performed using Duchon splines
  (see \code{\link{Duchon.spline}} for more information).
}

