% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/require2.R
\name{require2}
\alias{require2}
\title{Load packages with automatic installation if not installed}
\usage{
require2(pkgs)
}
\arguments{
\item{pkgs}{package names}
}
\value{
install packages
}
\description{
Load packages with automatic installation if not installed
}
\examples{
pkgs <- c("animation", "ggplot2")
require2(pkgs)
}
