% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSG-package.R
\docType{package}
\name{MSG-package}
\alias{MSG-package}
\alias{MSG}
\title{Modern Statistical Graphics}
\description{
Datasets and functions for the Chinese book ``Modern Statistical Graphics''.
}
\author{
Yihui Xie <\url{https://yihui.org}>
}
\keyword{package}
