% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msg.R
\name{msg}
\alias{msg}
\title{Plot a graph with a pre-installed R script}
\usage{
msg(fig = "3.6", show_code = TRUE, print_plot = TRUE, filter = 0)
}
\arguments{
\item{fig}{Character. The figure number or the R script name, which is given in the book.}

\item{show_code}{Logical. TRUE means the codes are shown in the console.}

\item{print_plot}{Logical. TRUE means the graph is printed.}

\item{filter}{Integer. The line numbers indicating which lines in the code are displayed (when positive) or hidden (when negative).}
}
\value{
A graph and the source code
}
\description{
Plot a graph with a pre-installed R script
}
\examples{
# msg('3.6') msg('ChinaPop')
}
